feat: 주식 기능 구현
- 학생 보유 주식 가져오기

S12P11D107-332 #done 학생 보유 주식 가져오기 완료

+++ frontend/Ttaenggrang/app/src/main/java/com/ladysparks/ttaenggrang/data/model/dto/StockTransactionDto.kt
@@ -3,7 +3,7 @@ package com.ladysparks.ttaenggrang.data.model.dto
 import com.google.gson.annotations.SerializedName
 
 data class StockTransactionDto(
-    val id: Long,
+    val id: Int,
     @SerializedName("share_count") val shareCount: Int,
     @SerializedName("trans_date") val transDate: String,
     @SerializedName("purchase_prc") val purchasePrc: Int,
@@ -11,8 +11,8 @@ data class StockTransactionDto(
     @SerializedName("return_amt") val returnAmt: Int,
     val transType: TransType,
     @SerializedName("owned_qty") val ownedQty: Int,
-    val studentId: Long,
-    val stockId: Long
+    val studentId: Int,
+    val stockId: Int
 )
 
 enum class TransType {
